Reorganize explorer
