+++ explorer/explorer.go
@@ -6,13 +6,9 @@
 package explorer
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"html/template"
-	"io"
 	"math"
-	"net/http"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -24,14 +20,12 @@ import (
 
 	"github.com/dcrdata/dcrdata/blockdata"
 	"github.com/dcrdata/dcrdata/db/dbtypes"
-	"github.com/decred/dcrd/chaincfg/chainhash"
 	"github.com/decred/dcrd/dcrjson"
 	"github.com/decred/dcrd/wire"
 	humanize "github.com/dustin/go-humanize"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	"github.com/rs/cors"
-	"golang.org/x/net/websocket"
 )
 
 const (
@@ -83,434 +77,6 @@ type explorerUI struct {
 	wsHub           *WebsocketHub
 	NewBlockDataMtx sync.RWMutex
 	NewBlockData    BlockBasic
-}
-
-func (exp *explorerUI) root(w http.ResponseWriter, r *http.Request) {
-	idx := exp.blockData.GetHeight()
-
-	height, err := strconv.Atoi(r.URL.Query().Get("height"))
-	if err != nil || height > idx {
-		height = idx
-	}
-
-	rows, err := strconv.Atoi(r.URL.Query().Get("rows"))
-	if err != nil || rows > maxExplorerRows || rows < minExplorerRows || height-rows < 0 {
-		rows = minExplorerRows
-	}
-	summaries := exp.blockData.GetExplorerBlocks(height, height-rows)
-	if summaries == nil {
-		log.Errorf("Unable to get blocks: height=%d&rows=%d", height, rows)
-		http.Redirect(w, r, "/error", http.StatusTemporaryRedirect)
-		return
-	}
-
-	str, err := templateExecToString(exp.templates[rootTemplateIndex], "explorer", struct {
-		Data      []*BlockBasic
-		BestBlock int
-	}{
-		summaries,
-		idx,
-	})
-
-	if err != nil {
-		log.Errorf("Template execute failure: %v", err)
-		http.Redirect(w, r, "/error", http.StatusTemporaryRedirect)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusOK)
-	io.WriteString(w, str)
-}
-
-func (exp *explorerUI) rootWebsocket(w http.ResponseWriter, r *http.Request) {
-	wsHandler := websocket.Handler(func(ws *websocket.Conn) {
-		// Create channel to signal updated data availability
-		updateSig := make(hubSpoke)
-		// register websocket client with our signal channel
-		exp.wsHub.RegisterClient(&updateSig)
-		// unregister (and close signal channel) before return
-		defer exp.wsHub.UnregisterClient(&updateSig)
-
-		requestLimit := 1 << 20
-		// set the max payload size to 1 MB
-		ws.MaxPayloadBytes = requestLimit
-
-		// Ticker for a regular ping
-		ticker := time.NewTicker(pingInterval)
-		defer ticker.Stop()
-
-		// Periodically ping clients over websocket connection
-		go func() {
-			for range ticker.C {
-				exp.wsHub.HubRelay <- sigPingAndUserCount
-			}
-		}()
-
-		// Start listening for websocket messages from client with raw
-		// transaction bytes (hex encoded) to decode or broadcast.
-		go func() {
-			for {
-				// Wait to receive a message on the websocket
-				msg := &WebSocketMessage{}
-				ws.SetReadDeadline(time.Now().Add(wsReadTimeout))
-				if err := websocket.JSON.Receive(ws, &msg); err != nil {
-					log.Warnf("websocket client receive error: %v", err)
-					return
-				}
-
-				// handle received message according to event ID
-				var webData WebSocketMessage
-				switch msg.EventId {
-				case "decodetx":
-					webData.EventId = msg.EventId + "Resp"
-					if len(msg.Message) > requestLimit {
-						log.Debug("Request size over limit")
-						webData.Message = "Request too large"
-						break
-					}
-					log.Debugf("Received decodetx signal for hex: %.40s...", msg.Message)
-					tx, err := exp.blockData.DecodeRawTransaction(msg.Message)
-					if err == nil {
-						message, err := json.MarshalIndent(tx, "", "    ")
-						if err != nil {
-							log.Warn("Invalid JSON message: ", err)
-							webData.Message = fmt.Sprintf("Error: Could not encode JSON message")
-							break
-						}
-						webData.Message = string(message)
-					} else {
-						log.Debugf("Could not decode raw tx")
-						webData.Message = fmt.Sprintf("Error: %v", err)
-					}
-				case "sendtx":
-					webData.EventId = msg.EventId + "Resp"
-					if len(msg.Message) > requestLimit {
-						log.Debugf("Request size over limit")
-						webData.Message = "Request too large"
-						break
-					}
-					log.Debugf("Received sendtx signal for hex: %.40s...", msg.Message)
-					txid, err := exp.blockData.SendRawTransaction(msg.Message)
-					if err != nil {
-						webData.Message = fmt.Sprintf("Error: %v", err)
-					} else {
-						webData.Message = fmt.Sprintf("Transaction sent: %s", txid)
-					}
-				case "ping":
-					log.Tracef("We've been pinged: %.40s...", msg.Message)
-					continue
-				default:
-					log.Warnf("Unrecognized event ID: %v", msg.EventId)
-					continue
-				}
-
-				// send the response back on the websocket
-				ws.SetWriteDeadline(time.Now().Add(wsWriteTimeout))
-				if err := websocket.JSON.Send(ws, webData); err != nil {
-					log.Debugf("Failed to encode WebSocketMessage %s: %v",
-						webData.EventId, err)
-					// If the send failed, the client is probably gone, so close
-					// the connection and quit.
-					return
-				}
-			}
-		}()
-
-		// Ping and block update loop (send only)
-	loop:
-		for {
-			// Wait for signal from the hub to update
-			select {
-			case sig, ok := <-updateSig:
-				// Check if the update channel was closed. Either the websocket
-				// hub will do it after unregistering the client, or forcibly in
-				// response to (http.CloseNotifier).CloseNotify() and only then if
-				// the hub has somehow lost track of the client.
-				if !ok {
-					//ws.WriteClose(1)
-					exp.wsHub.UnregisterClient(&updateSig)
-					break loop
-				}
-
-				if _, ok = eventIDs[sig]; !ok {
-					break loop
-				}
-
-				log.Tracef("signaling client: %p", &updateSig)
-
-				// Write block data to websocket client
-				exp.NewBlockDataMtx.RLock()
-				webData := WebSocketMessage{
-					EventId: eventIDs[sig],
-				}
-				buff := new(bytes.Buffer)
-				enc := json.NewEncoder(buff)
-				switch sig {
-				case sigNewBlock:
-					enc.Encode(WebsocketBlock{exp.NewBlockData})
-					webData.Message = buff.String()
-				case sigPingAndUserCount:
-					// ping and send user count
-					webData.Message = strconv.Itoa(exp.wsHub.NumClients())
-				}
-
-				ws.SetWriteDeadline(time.Now().Add(wsWriteTimeout))
-				err := websocket.JSON.Send(ws, webData)
-				exp.NewBlockDataMtx.RUnlock()
-				if err != nil {
-					log.Debugf("Failed to encode WebSocketMessage %v: %v", sig, err)
-					// If the send failed, the client is probably gone, so close
-					// the connection and quit.
-					return
-				}
-			case <-exp.wsHub.quitWSHandler:
-				break loop
-			}
-		}
-	})
-
-	wsHandler.ServeHTTP(w, r)
-}
-
-func (exp *explorerUI) blockPage(w http.ResponseWriter, r *http.Request) {
-	hash := getBlockHashCtx(r)
-
-	data := exp.blockData.GetExplorerBlock(hash)
-	if data == nil {
-		log.Errorf("Unable to get block %s", hash)
-		http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-		return
-	}
-
-	pageData := struct {
-		Data          *BlockInfo
-		ConfirmHeight int64
-	}{
-		data,
-		exp.NewBlockData.Height - data.Confirmations,
-	}
-	str, err := templateExecToString(exp.templates[blockTemplateIndex], "block", pageData)
-	if err != nil {
-		log.Errorf("Template execute failure: %v", err)
-		http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusOK)
-	io.WriteString(w, str)
-}
-
-func (exp *explorerUI) txPage(w http.ResponseWriter, r *http.Request) {
-	// attempt to get tx hash string from URL path
-	hash, ok := r.Context().Value(ctxTxHash).(string)
-	if !ok {
-		log.Trace("txid not set")
-		http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-		return
-	}
-	tx := exp.blockData.GetExplorerTx(hash)
-	if tx == nil {
-		log.Errorf("Unable to get transaction %s", hash)
-		http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-		return
-	}
-	if !exp.liteMode {
-		// For each output of this transaction, look up any spending transactions,
-		// and the index of the spending transaction input.
-		spendingTxHashes, spendingTxVinInds, voutInds, err := exp.explorerSource.SpendingTransactions(hash)
-		if err != nil {
-			log.Errorf("Unable to retrieve spending transactions for %s: %v", hash, err)
-			http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-			return
-		}
-		for i, vout := range voutInds {
-			if int(vout) >= len(tx.SpendingTxns) {
-				log.Errorf("Invalid spending transaction data (%s:%d)", hash, vout)
-				continue
-			}
-			tx.SpendingTxns[vout] = TxInID{
-				Hash:  spendingTxHashes[i],
-				Index: spendingTxVinInds[i],
-			}
-		}
-	}
-
-	pageData := struct {
-		Data          *TxInfo
-		ConfirmHeight int64
-	}{
-		tx,
-		exp.NewBlockData.Height - tx.Confirmations,
-	}
-
-	str, err := templateExecToString(exp.templates[txTemplateIndex], "tx", pageData)
-	if err != nil {
-		log.Errorf("Template execute failure: %v", err)
-		http.Redirect(w, r, "/error/"+hash, http.StatusTemporaryRedirect)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusOK)
-	io.WriteString(w, str)
-}
-
-func (exp *explorerUI) addressPage(w http.ResponseWriter, r *http.Request) {
-	// Get the address URL parameter, which should be set in the request context
-	// by the addressPathCtx middleware.
-	address, ok := r.Context().Value(ctxAddress).(string)
-	if !ok {
-		log.Trace("address not set")
-		http.Redirect(w, r, "/error/"+address, http.StatusTemporaryRedirect)
-		return
-	}
-
-	// Number of outputs for the address to query the database for. The URL
-	// query parameter "n" is used to specify the limit (e.g. "?n=20").
-	limitN, err := strconv.ParseInt(r.URL.Query().Get("n"), 10, 64)
-	if err != nil || limitN < 0 {
-		limitN = defaultAddressRows
-	} else if limitN > maxAddressRows {
-		log.Warnf("addressPage: requested up to %d address rows, "+
-			"limiting to %d", limitN, maxAddressRows)
-		limitN = maxAddressRows
-	}
-
-	// Number of outputs to skip (OFFSET in database query). For UX reasons, the
-	// "start" URL query parameter is used.
-	offsetAddrOuts, err := strconv.ParseInt(r.URL.Query().Get("start"), 10, 64)
-	if err != nil || offsetAddrOuts < 0 {
-		offsetAddrOuts = 0
-	}
-
-	var addrData *AddressInfo
-	if exp.liteMode {
-		addrData = exp.blockData.GetExplorerAddress(address, limitN, offsetAddrOuts)
-		if addrData == nil {
-			log.Errorf("Unable to get address %s", address)
-			http.Redirect(w, r, "/error/"+address, http.StatusTemporaryRedirect)
-			return
-		}
-	} else {
-		// Get addresses table rows for the address
-		addrHist, balance, errH := exp.explorerSource.AddressHistory(
-			address, limitN, offsetAddrOuts)
-		if errH != nil {
-			log.Errorf("Unable to get address %s history: %v", address, errH)
-			http.Redirect(w, r, "/error/"+address, http.StatusTemporaryRedirect)
-			return
-		}
-
-		// Generate AddressInfo skeleton from the address table rows
-		addrData = ReduceAddressHistory(addrHist)
-		if addrData == nil {
-			log.Debugf("empty address history (%s): n=%d&start=%d", address, limitN, offsetAddrOuts)
-			http.Redirect(w, r, "/error/"+address, http.StatusTemporaryRedirect)
-			return
-		}
-		addrData.Limit, addrData.Offset = limitN, offsetAddrOuts
-		addrData.KnownFundingTxns = balance.NumSpent + balance.NumUnspent
-		addrData.Balance = balance
-		addrData.Path = r.URL.Path
-		// still need []*AddressTx filled out and NumUnconfirmed
-
-		// Query database for transaction details
-		err = exp.explorerSource.FillAddressTransactions(addrData)
-		if err != nil {
-			log.Errorf("Unable to fill address %s transactions: %v", address, err)
-			http.Redirect(w, r, "/error/"+address, http.StatusTemporaryRedirect)
-			return
-		}
-	}
-
-	confirmHeights := make([]int64, len(addrData.Transactions))
-	for i, v := range addrData.Transactions {
-		confirmHeights[i] = exp.NewBlockData.Height - int64(v.Confirmations)
-	}
-	pageData := struct {
-		Data          *AddressInfo
-		ConfirmHeight []int64
-	}{
-		addrData,
-		confirmHeights,
-	}
-
-	str, err := templateExecToString(exp.templates[addressTemplateIndex], "address", pageData)
-	if err != nil {
-		log.Errorf("Template execute failure: %v", err)
-		http.Redirect(w, r, "/error", http.StatusTemporaryRedirect)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusOK)
-	io.WriteString(w, str)
-}
-
-func (exp *explorerUI) decodeTxPage(w http.ResponseWriter, r *http.Request) {
-	str, err := templateExecToString(exp.templates[decodeTxTemplateIndex], "rawtx", nil)
-	if err != nil {
-		log.Errorf("Template execute failure: %v", err)
-		http.Redirect(w, r, "/error", http.StatusTemporaryRedirect)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusOK)
-	io.WriteString(w, str)
-}
-
-// search implements a primitive search algorithm by checking if the value in
-// question is a block index, block hash, address hash or transaction hash and
-// redirects to the appropriate page or displays an error
-func (exp *explorerUI) search(w http.ResponseWriter, r *http.Request) {
-	searchStr, ok := r.Context().Value(ctxSearch).(string)
-	if !ok {
-		log.Trace("search parameter missing")
-		http.Redirect(w, r, "/error/", http.StatusTemporaryRedirect)
-		return
-	}
-
-	// Attempt to get a block hash by calling GetBlockHash to see if the value
-	// is a block index and then redirect to the block page if it is
-	idx, err := strconv.ParseInt(searchStr, 10, 0)
-	if err == nil {
-		_, err = exp.blockData.GetBlockHash(idx)
-		if err == nil {
-			http.Redirect(w, r, "/explorer/block/"+searchStr, http.StatusPermanentRedirect)
-			return
-		}
-	}
-
-	// Call GetExplorerAddress to see if the value is an address hash and
-	// then redirect to the address page if it is
-	address := exp.blockData.GetExplorerAddress(searchStr, 1, 0)
-	if address != nil {
-		http.Redirect(w, r, "/explorer/address/"+searchStr, http.StatusPermanentRedirect)
-		return
-	}
-
-	// Check if the value is a valid hash
-	if _, err = chainhash.NewHashFromStr(searchStr); err != nil {
-		http.Redirect(w, r, "/error/"+searchStr, http.StatusTemporaryRedirect)
-		return
-	}
-
-	// Attempt to get a block index by calling GetBlockHeight to see if the
-	// value is a block hash and then redirect to the block page if it is
-	_, err = exp.blockData.GetBlockHeight(searchStr)
-	if err == nil {
-		http.Redirect(w, r, "/explorer/block/"+searchStr, http.StatusPermanentRedirect)
-		return
-	}
-
-	// Call GetExplorerTx to see if the value is a transaction hash and then
-	// redirect to the tx page if it is
-	tx := exp.blockData.GetExplorerTx(searchStr)
-	if tx != nil {
-		http.Redirect(w, r, "/explorer/tx/"+searchStr, http.StatusPermanentRedirect)
-		return
-	}
-
-	// Display an error since searchStr is not a block index, block hash, address hash or transaction hash
-	http.Redirect(w, r, "/error/"+searchStr, http.StatusTemporaryRedirect)
 }
 
 func (exp *explorerUI) reloadTemplates() error {
